+++ app/models/offer.rb
@@ -47,6 +47,8 @@ class Offer < ApplicationRecord
   before_validation :set_internal
   before_validation :set_oms_details
   before_validation :sanitize_oms_params
+  before_validation :set_iid
+  after_initialize :set_iid
 
   has_many :bundle_offers
   has_many :bundles, through: :bundle_offers, dependent: :destroy
@@ -56,17 +58,19 @@ class Offer < ApplicationRecord
   belongs_to :offer_type, class_name: "Vocabulary::ServiceCategory", optional: true
   belongs_to :offer_subtype, class_name: "Vocabulary::ServiceCategory", optional: true
 
-  validate :set_iid, on: :create
   validates :service, presence: true
   validates :iid, presence: true, numericality: true
   validates :order_url, mp_url: true, if: :order_url?
 
   validate :primary_oms_exists?, if: -> { primary_oms_id.present? }
-  validate :proper_oms?, if: -> { primary_oms.present? }
-  validates :oms_params, absence: true, if: -> { current_oms.blank? }
-  validate :check_oms_params, if: -> { current_oms.present? }
   validate :check_main_bundles, if: -> { draft? }
-  validate :same_order_type_as_in_service, if: -> { service&.order_type.present? }
+
+  with_options unless: :draft? do
+    validate :proper_oms?, if: -> { primary_oms.present? }
+    validates :oms_params, absence: true, if: -> { current_oms.blank? }
+    validate :check_oms_params, if: -> { current_oms.present? }
+    validate :same_order_type_as_in_service, if: -> { service&.order_type.present? }
+  end
 
   before_destroy :check_main_bundles