+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Added
 - Environmental variable to hide recommendation panel (`SHOW_RECOMMENDATION_PANEL`) (@maria-j-k)
 - Unit to the constant offer parameter (@gorreck888)
+- `Save as draft` and `Duplicate` features for offer form (@maria-j-k) 
 
 ### Changed
 - MarketplaceLocation renamed to ResearchActivity (@maria-j-k)